Moved hardware and environments  into folders
