+++ src/rattlesnake/hardware/exodus_modal_solution_hardware.py
@@ -30,8 +30,8 @@ import netCDF4
 import numpy as np
 import scipy.signal as signal
 
-from rattlesnake.components.abstract_hardware import HardwareAcquisition, HardwareOutput
-from rattlesnake.components.utilities import Channel, DataAcquisitionParameters, flush_queue
+from rattlesnake.hardware.abstract_hardware import HardwareAcquisition, HardwareOutput
+from rattlesnake.utilities import Channel, DataAcquisitionParameters, flush_queue
 
 DEBUG = False
