+++ src/rattlesnake/hardware/lanxi_hardware_multiprocessing.py
@@ -30,9 +30,9 @@ from typing import List
 import numpy as np
 import requests
 
-from rattlesnake.components.abstract_hardware import HardwareAcquisition, HardwareOutput
-from rattlesnake.components.lanxi_stream import OpenapiHeader, OpenapiMessage
-from rattlesnake.components.utilities import Channel, DataAcquisitionParameters
+from rattlesnake.hardware.abstract_hardware import HardwareAcquisition, HardwareOutput
+from rattlesnake.hardware.lanxi_stream import OpenapiHeader, OpenapiMessage
+from rattlesnake.utilities import Channel, DataAcquisitionParameters
 
 OUTPUT_RATE = 131072
 LANXI_STATE_TIMEOUT = 255.0
