+++ src/rattlesnake/hardware/nidaqmx_hardware_multitask.py
@@ -31,8 +31,8 @@ import nidaqmx.stream_readers as ni_read
 import nidaqmx.stream_writers as ni_write
 import numpy as np
 
-from rattlesnake.components.abstract_hardware import HardwareAcquisition, HardwareOutput
-from rattlesnake.components.utilities import Channel, DataAcquisitionParameters
+from rattlesnake.hardware.abstract_hardware import HardwareAcquisition, HardwareOutput
+from rattlesnake.utilities import Channel, DataAcquisitionParameters
 
 BUFFER_SIZE_FACTOR = 3
