+++ src/rattlesnake/hardware/state_space_virtual_hardware.py
@@ -31,8 +31,8 @@ import numpy as np
 import scipy.signal as signal
 from scipy.io import loadmat
 
-from rattlesnake.components.abstract_hardware import HardwareAcquisition, HardwareOutput
-from rattlesnake.components.utilities import Channel, DataAcquisitionParameters, flush_queue
+from rattlesnake.hardware.abstract_hardware import HardwareAcquisition, HardwareOutput
+from rattlesnake.utilities import Channel, DataAcquisitionParameters, flush_queue
 
 
 class StateSpaceAcquisition(HardwareAcquisition):
