+++ src/rattlesnake/process/abstract_sysid_data_analysis.py
@@ -27,8 +27,8 @@ import multiprocessing as mp
 from enum import Enum
 
 from rattlesnake.process.abstract_message_process import AbstractMessageProcess
-from rattlesnake.components.abstract_sysid_environment import AbstractSysIdMetadata
-from rattlesnake.components.utilities import VerboseMessageQueue, flush_queue
+from rattlesnake.environment.abstract_sysid_environment import AbstractSysIdMetadata
+from rattlesnake.utilities import VerboseMessageQueue, flush_queue
 
 
 class SysIDDataAnalysisCommands(Enum):
