+++ src/rattlesnake/process/output.py
@@ -28,7 +28,7 @@ import multiprocessing.sharedctypes  # pylint: disable=unused-import
 import numpy as np
 
 from rattlesnake.process.abstract_message_process import AbstractMessageProcess
-from rattlesnake.components.utilities import GlobalCommands, QueueContainer, flush_queue, rms_time
+from rattlesnake.utilities import GlobalCommands, QueueContainer, flush_queue, rms_time
 
 TASK_NAME = "Output"
 
@@ -153,42 +153,42 @@ class OutputProcess(AbstractMessageProcess):
         if self.hardware is not None:
             self.hardware.close()
         if data_acquisition_parameters.hardware == 0:
-            from ..components.nidaqmx_hardware_multitask import NIDAQmxOutput
+            from ..hardware.nidaqmx_hardware_multitask import NIDAQmxOutput
 
             self.hardware = NIDAQmxOutput(
                 data_acquisition_parameters.extra_parameters["task_trigger"],
                 data_acquisition_parameters.extra_parameters["task_trigger_output_channel"],
             )
         elif data_acquisition_parameters.hardware == 1:
-            from ..components.lanxi_hardware_multiprocessing import LanXIOutput
+            from ..hardware.lanxi_hardware_multiprocessing import LanXIOutput
 
             self.hardware = LanXIOutput(
                 data_acquisition_parameters.extra_parameters["maximum_acquisition_processes"]
             )
         elif data_acquisition_parameters.hardware == 2:
-            from ..components.data_physics_hardware import DataPhysicsOutput
+            from ..hardware.data_physics_hardware import DataPhysicsOutput
 
             self.hardware = DataPhysicsOutput(self.queue_container.single_process_hardware_queue)
         elif data_acquisition_parameters.hardware == 3:
-            from ..components.data_physics_dp900_hardware import DataPhysicsDP900Output
+            from ..hardware.data_physics_dp900_hardware import DataPhysicsDP900Output
 
             self.hardware = DataPhysicsDP900Output(
                 self.queue_container.single_process_hardware_queue,
             )
         elif data_acquisition_parameters.hardware == 4:
-            from ..components.exodus_modal_solution_hardware import ExodusOutput
+            from ..hardware.exodus_modal_solution_hardware import ExodusOutput
 
             self.hardware = ExodusOutput(self.queue_container.single_process_hardware_queue)
         elif data_acquisition_parameters.hardware == 5:
-            from ..components.state_space_virtual_hardware import StateSpaceOutput
+            from ..hardware.state_space_virtual_hardware import StateSpaceOutput
 
             self.hardware = StateSpaceOutput(self.queue_container.single_process_hardware_queue)
         elif data_acquisition_parameters.hardware == 6:
-            from ..components.sdynpy_system_virtual_hardware import SDynPySystemOutput
+            from ..hardware.sdynpy_system_virtual_hardware import SDynPySystemOutput
 
             self.hardware = SDynPySystemOutput(self.queue_container.single_process_hardware_queue)
         elif data_acquisition_parameters.hardware == 7:
-            from ..components.sdynpy_frf_virtual_hardware import SDynPyFRFOutput
+            from ..hardware.sdynpy_frf_virtual_hardware import SDynPyFRFOutput
 
             self.hardware = SDynPyFRFOutput(self.queue_container.single_process_hardware_queue)
         else:
