+++ src/rattlesnake/process/signal_generation_process.py
@@ -27,7 +27,7 @@ import numpy as np
 
 from rattlesnake.process.abstract_message_process import AbstractMessageProcess
 from rattlesnake.process.signal_generation import SignalGenerator
-from rattlesnake.components.utilities import VerboseMessageQueue, flush_queue, rms_time
+from rattlesnake.utilities import VerboseMessageQueue, flush_queue, rms_time
 
 TEST_LEVEL_THRESHOLD = 1.01
