+++ tests/test_abstract_environment.py
@@ -3,14 +3,15 @@ from unittest import mock
 
 import pytest
 from functions.common_functions import fake_time
+
 # from PyQt5 import QtWidgets  # comment out unused import
 
-from rattlesnake.components.abstract_environment import (
+from rattlesnake.environment.abstract_environment import (
     AbstractEnvironment,
     AbstractMetadata,
     AbstractUI,
 )
-from rattlesnake.components.utilities import GlobalCommands, VerboseMessageQueue
+from rattlesnake.utilities import GlobalCommands, VerboseMessageQueue
 
 
 # Initialize log_file_queue that verbose queues will use
@@ -72,9 +73,7 @@ class DummyAbstractMetadata(AbstractMetadata):
 
 # Create a dummy AbstractUI class
 class DummyAbstractUI(AbstractUI):
-    def __init__(
-        self, log_file_queue, environment_command_queue, controller_command_queue
-    ):
+    def __init__(self, log_file_queue, environment_command_queue, controller_command_queue):
         super().__init__(
             "Environment Name",
             environment_command_queue,
@@ -135,9 +134,7 @@ class DummyAbstractEnvironment(AbstractEnvironment):
         )
 
     def initialize_data_acquisition_parameters(self, data_acquisition_parameters):
-        return super().initialize_data_acquisition_parameters(
-            data_acquisition_parameters
-        )
+        return super().initialize_data_acquisition_parameters(data_acquisition_parameters)
 
     def initialize_environment_test_parameters(self, environment_parameters):
         return super().initialize_environment_test_parameters(environment_parameters)
@@ -149,9 +146,7 @@ class DummyAbstractEnvironment(AbstractEnvironment):
 # Initialize the AbstractUI class
 @pytest.fixture
 def abstract_ui(log_file_queue, environment_command_queue, controller_command_queue):
-    return DummyAbstractUI(
-        log_file_queue, environment_command_queue, controller_command_queue
-    )
+    return DummyAbstractUI(log_file_queue, environment_command_queue, controller_command_queue)
 
 
 # Initialize the AbstractEnvironment class
@@ -247,9 +242,7 @@ def test_abstract_environment_init(
     # Test the environment_command_queue property
     assert abstract_environment.environment_command_queue == environment_command_queue
     # Test the controller_communication_queue property
-    assert (
-        abstract_environment.controller_communication_queue == controller_command_queue
-    )
+    assert abstract_environment.controller_communication_queue == controller_command_queue
     # Test the log_file_queue property
     assert abstract_environment.log_file_queue == log_file_queue
     # Test the gui_update_queue property
@@ -281,9 +274,7 @@ def test_abstract_environment_log(mock_time, mock_put, abstract_environment):
     abstract_environment.log(message)
 
     # Test if the correct string was writtin to log_file_queue
-    mock_put.assert_called_with(
-        "{:}: {:} -- {:}\n".format("Datetime", "Environment Name", message)
-    )
+    mock_put.assert_called_with("{:}: {:} -- {:}\n".format("Datetime", "Environment Name", message))
 
 
 # Test the AbstractEnvironment quit function
