+++ tests/test_abstract_hardware.py
@@ -1,4 +1,4 @@
-from rattlesnake.components.abstract_hardware import HardwareAcquisition, HardwareOutput
+from rattlesnake.hardware.abstract_hardware import HardwareAcquisition, HardwareOutput
 from unittest import mock
 import numpy as np
 import pytest
@@ -10,25 +10,25 @@ class DummyHardwareAcquisition(HardwareAcquisition):
 
     def set_up_data_acquisition_parameters_and_channels(self, test_data, channel_data):
         return super().set_up_data_acquisition_parameters_and_channels(test_data, channel_data)
-    
+
     def start(self):
         return super().start()
-    
+
     def read(self):
         return super().read()
-    
+
     def read_remaining(self):
         return super().read_remaining()
-    
+
     def stop(self):
         return super().stop()
-    
+
     def close(self):
         return super().close()
-    
+
     def get_acquisition_delay(self):
         return super().get_acquisition_delay()
-    
+
 
 class DummyHardwareOutput(HardwareOutput):
     def __init__(self):
@@ -36,22 +36,22 @@ class DummyHardwareOutput(HardwareOutput):
 
     def set_up_data_output_parameters_and_channels(self, test_data, channel_data):
         return super().set_up_data_output_parameters_and_channels(test_data, channel_data)
-    
+
     def start(self):
         return super().start()
-    
+
     def write(self, data):
         return super().write(data)
-    
+
     def stop(self):
         return super().stop()
-    
+
     def close(self):
         return super().close()
-    
+
     def ready_for_new_output(self):
         return super().ready_for_new_output()
-    
+
 
 def test_hardware_acquisition_init():
     hardware_acquistion = DummyHardwareAcquisition()
