+++ tests/test_sdynpy_system_virtual_hardware.py
@@ -1,10 +1,5 @@
-from rattlesnake.components.sdynpy_system_virtual_hardware import SDynPySystemAcquisition
+from rattlesnake.hardware.sdynpy_system_virtual_hardware import SDynPySystemAcquisition
 from unittest import mock
 import multiprocessing as mp
 import pytest
 import numpy as np
-
-
-
-
-
