+++ tests/test_spectral_processing.py
@@ -5,7 +5,7 @@ from unittest import mock
 import numpy as np
 import pytest
 
-from rattlesnake.components.spectral_processing import (
+from rattlesnake.process.spectral_processing import (
     AveragingTypes,
     Estimator,
     SpectralProcessingCommands,
@@ -13,7 +13,7 @@ from rattlesnake.components.spectral_processing import (
     SpectralProcessingProcess,
     spectral_processing_process,
 )
-from rattlesnake.components.utilities import GlobalCommands, VerboseMessageQueue
+from rattlesnake.utilities import GlobalCommands, VerboseMessageQueue
 
 
 # Create log_file_queue
@@ -175,9 +175,7 @@ def test_spectral_processing_clear_spectral_processing(spectral_processing_proce
     )
 
 
-@mock.patch(
-    "rattlesnake.components.spectral_processing.SpectralProcessingProcess.command_queue"
-)
+@mock.patch("rattlesnake.components.spectral_processing.SpectralProcessingProcess.command_queue")
 @mock.patch("rattlesnake.components.spectral_processing.flush_queue")
 def test_spectral_processing_stop_spectral_processing(
     mock_flush, mock_command, spectral_processing_process_obj
@@ -201,9 +199,7 @@ def test_spectral_processing_stop_spectral_processing(
 
 # Test the spectral_processing_process function
 # Prevent run while loop from starting
-@mock.patch(
-    "rattlesnake.components.abstract_message_process.AbstractMessageProcess.run"
-)
+@mock.patch("rattlesnake.components.abstract_message_process.AbstractMessageProcess.run")
 def test_spectral_processing_process(mock_run, log_file_queue):
     spectral_processing_process(
         "Environment Name",
