+++ tests/test_streaming.py
@@ -1,5 +1,5 @@
 from rattlesnake.process.streaming import StreamingProcess, streaming_process
-from rattlesnake.components.utilities import (
+from rattlesnake.utilities import (
     VerboseMessageQueue,
     QueueContainer,
     Channel,
