+++ tests/test_ui_utilities.py
@@ -39,7 +39,7 @@ from rattlesnake.components.ui_utilities import (
     multiline_plotter,
     save_combined_environments_profile_template,
 )
-from rattlesnake.components.utilities import Channel, DataAcquisitionParameters
+from rattlesnake.utilities import Channel, DataAcquisitionParameters
 
 
 # Initialize an QApplication which is required for all other widgets
@@ -161,9 +161,7 @@ def data_acquisition_parameters(channel_list):
 
 @pytest.fixture
 def channel_monitor(app, data_acquisition_parameters):
-    with mock.patch(
-        "rattlesnake.components.ui_utilities.QtWidgets.QDialog.show"
-    ) as mock_show:
+    with mock.patch("rattlesnake.components.ui_utilities.QtWidgets.QDialog.show") as mock_show:
         channel_monitor = ChannelMonitor(None, data_acquisition_parameters)
         app.addWidget(channel_monitor)
